+++ NEWS.md
@@ -4,6 +4,7 @@
 * Show check outputs in console
 * Ask before overwriting everything
 * Check Description Title and description fields
+* Check if folder name is correct package name
 
 # v0.2.2